pyproject.toml
[tool.poetry]
name = "vector-embedding-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "vector_embedding_server"}]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.30.2"
torch = {version = "^2.0.1", source = "torch"}
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
httpx = "^0.24.1"
fastapi-security = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.0"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.2"
openai = "^0.27.8"
requests = "^2.31.0"
flake8 = "^6.0.0"
types-python-jose = "^3.3.4.7"
types-passlib = "^1.7.7.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"



[tool.isort]
profile = "black"
multi_line_output = 3